common2 와 merge 후 commit & push 헤더, 게시판 리스트 나오는것까지 확인함(신철희 12/19 12:12)

+++ src/main/java/com/inseoul/board/controller/CommentController.java
@@ -26,7 +26,7 @@ public class CommentController {
     @PostMapping("/write")
     public QryResult write(
             @RequestParam("post_id") Long postId,
-            @RequestParam("user_id")Long userId,
+            @RequestParam("user_id") int userId,
             String content
     ){
         return commentService.write(postId, userId, content);
